Added GameCameraSystem
Simple Camera system that can focus on objects and can contain points of interest that behave acoording to Point Of interest Controllers or PoiController's with parameters such as weight.duration,number of activations.

Adding a focus object is as easy as calling GameCameraSystem.addFocus(focusobject),And adding a point of interest is as easy as adding a PoiController to an existing GameObject on the same GameLayer as the CameraSystem. Focus Objects and Points of Interests Behave differently and should be used accoordingly(Points of interest :for things you want to draw attention to but do not wirh to have as the main focus, Focus Object: the main focus of the camera, these objects will always try to be kept in frame ).

Also added some Screen shake functionality to this System.