Added ParticleSystem & ActionEffects
Particle system that batches effects. Particle effects are pooled to prevent garbage collection. Adding a ParticleController to a GameObject renders it on the gamelayer of the system & gameobject.

The controller has duration scale and file fields as well as a post -pre draw.

Since all effects are batchhed Additive blended effects will always be rendered before all others.

--Additions
--- added a spiral effect
---added a shake effect

--Fixes
---GameStates are now Stage Groups
---addGameObjectController shortened to addController