InGame DevConsole and other cool stuff
Finally added an extensible dev console from usable within games. The console must be instantiated by the developer. The goal is to remove the instantiation from launch builds if that is desired.

Added some shader optimizations and some quality of life changes to GameObjects as well. They can now support an infinite amount of child depth and they no longer flush the batch. This was an oversight that is now fixed.