added the ability to use an Asset tmx base
Lets you define a preload tmx to use to load assets instead of specifying assets to load in the preload layer. This is relatively useful for instances where you have several of the same Tmx files(for example different levels) that use the same assets. So you can make changes to the assets tmx and they will be carried to all who use it as a preload.