fixed a major bug
fixed a bug that was causing touch to be disabled when switching screens. What was happenign is that when we pushed a screen on top of another it would disable the one underneath to avoid picking up touch events but was not being reset again when switching back without poping the top state.